#37 refactor: 외부 요청 처리를 ExternalApiService에서 수행하도록 수정
